+++ src/commons/utils/password.utils.ts
@@ -1,0 +1,61 @@
+import * as crypto from 'crypto';
+
+export class PasswordUtils {
+    public static DEFAULTCHARSET = 'abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ';
+    public static DEFAULTNUMBERCHARSET = '0123456789';
+    public static DEFAULTSPECIALCHARSET = '_@-()<>:';
+
+    public static createNewPassword(max: number, chars: string, numberChars: string, specialChars: string): string {
+        if (chars.length < 2 && numberChars.length < 2 && specialChars.length < 2) {
+            throw new Error('minimum one of the charsets must have length >= 2');
+        }
+
+        const a = [];
+        while (a.length < max) {
+            let charset: string;
+            // TODO check charsetlength spread
+            const charSetNumber = Math.floor(Math.random() * 3);
+            switch (charSetNumber) {
+                case 0:
+                    charset = chars;
+                    break;
+                case 1:
+                    charset = numberChars;
+                    break;
+                case 2:
+                    charset = specialChars;
+                    break;
+                default:
+                    throw new Error('unknown charset-index:' + charSetNumber)
+            }
+
+            if (charset.length < 2) {
+                continue;
+            }
+
+            a.push(charset[Math.floor(Math.random() * charset.length)]);
+        }
+
+        return a.join('');
+    }
+
+    public static createNewDefaultPassword(max: number): string {
+        return this.createNewPassword(max, this.DEFAULTCHARSET, this.DEFAULTNUMBERCHARSET, this.DEFAULTSPECIALCHARSET);
+    }
+
+    public static createSolrPasswordHash(pwd: string): Promise<string> {
+        return new Promise<string>((accept, reject) => {
+            const salt: Buffer = crypto.randomBytes(32);
+            const saltUtf = salt.toString('utf8');
+            const saltBase64 = salt.toString('base64');
+            crypto.pbkdf2(pwd, saltUtf, 1000, 32, 'sha256', (error, hash: Buffer) => {
+                if (error) {
+                    reject(error);
+                } else {
+                    accept(hash.toString('base64') + ' ' + saltBase64);
+                }
+            });
+        });
+    }
+
+}