+++ src/facetcache-commons/model/sqlite3-facetcache.adapter.ts
@@ -215,6 +215,6 @@ export class Sqlite3FacetCacheAdapter implements FacetCacheAdapter {
     }
 
     protected extractSqlFileOnScriptPath(sqlFile: string, splitter: string): string[] {
-        return DatabaseService.extractSqlFileOnScriptPath(this.sqlScriptPath, splitter);
+        return DatabaseService.extractSqlFileOnScriptPath(this.sqlScriptPath + '/' + sqlFile, splitter);
     }
 }